+++ lib/holdify/hold.rb
@@ -3,15 +3,15 @@
 module Holdify
   # The *_hold statement (Assertion/Expectation)
   class Hold
-    attr_reader :forced
+    attr_reader :forced, :store
 
     def initialize(test)
       @test    = test
       @path,   = test.method(test.name).source_location
       @store   = Holdify.stores(@path)
       @session = Hash.new { |h, k| h[k] = [] } # { line => [values] }
-      @forced  = []                            # [ line ]
-      @added   = []                            # [ line ]
+      @forced  = []                            # [ lines ]
+      @added   = []                            # [ lines ]
     end
 
     def call(actual, force: false)
@@ -41,9 +41,12 @@ module Holdify
       @session.each { |line, values| @store.set(line, values) }
     end
 
+    # The index of the current test/value
+    def current_index(line) = @session[line].size - 1
+
     # Find the location in the test that triggered the hold
     def find_location
-      caller_locations.find do |location|
+      caller_locations(2).find do |location|
         next unless location.path == @path
 
         label = location.base_label