+++ test/holdify_test.rb
@@ -30,47 +30,6 @@ describe 'holdify' do
       FileUtils.rm_f("#{path}#{Holdify::CONFIG[:ext]}")
       Holdify.stores.delete(path)
     end
-
-    describe 'pretty' do
-      after { Holdify.pretty = nil }
-
-      it 'defaults to auto-detection' do
-        Holdify.pretty = nil
-        $stdout.stub :tty?, true do
-          Holdify.stub :system, true do
-            assert Holdify.pretty
-          end
-          Holdify.stub :system, false do
-            Holdify.pretty = nil
-            refute Holdify.pretty
-          end
-        end
-
-        $stdout.stub :tty?, false do
-          Holdify.pretty = nil
-          refute Holdify.pretty
-        end
-      end
-
-      it 'respects NO_COLOR' do
-        Holdify.pretty = nil
-        ENV['NO_COLOR'] = '1'
-        $stdout.stub :tty?, true do
-          Holdify.stub :system, true do
-            refute Holdify.pretty
-          end
-        end
-      ensure
-        ENV.delete('NO_COLOR')
-      end
-
-      it 'can be forced' do
-        Holdify.pretty = true
-        assert Holdify.pretty
-        Holdify.pretty = false
-        refute Holdify.pretty
-      end
-    end
   end
 
   describe 'persist_all!' do